requirements.txt
# Django Framework
Django>=5.2.0,<6.0.0
djangorestframework>=3.16.0,<4.0.0

# Llamaindex for document processing
llama-index>=0.13.0
llama-index-core>=0.13.0
llama-index-readers-file>=0.4.0

# Document processing support
pypdf>=3.0.0
python-docx>=1.0.0
pillow>=10.3.0

# API and utilities
python-multipart>=0.0.18
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.0

# CORS support
django-cors-headers>=4.0.0
